refactor : WeightFeedback import
